src/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
decouple==0.0.7
dj-database-url==0.5.0
Django==3.1.4
django-oauth-toolkit==1.3.3
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-decouple==3.3
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
